requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.35.0
diffusers>=0.21.0
accelerate>=0.24.0
safetensors>=0.4.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Discord Bot
discord.py>=2.3.0
python-dotenv>=1.0.0

# ComfyUI Integration
aiohttp>=3.9.0
websockets>=12.0
requests>=2.31.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0

# Qwen Support
tiktoken>=0.5.0

# Optional: For better performance
xformers>=0.0.22  # Optional, for memory efficiency

